refactor: create top-tab for token assets
