+++ packages/reactnative/src/screens/Dashboard/Tab/DebugContracts/index.tsx
@@ -1,0 +1,1 @@
+export { default as DebugContracts } from "./DebugContracts";
